Improve schema docs and move all SKU variant related types under the same `skuVariants` property
