+++ packages/api/src/platforms/vtex/index.ts
@@ -20,6 +20,7 @@ import { Query } from './resolvers/query'
 import { StoreReview } from './resolvers/review'
 import { StoreSearchResult } from './resolvers/searchResult'
 import { StoreSeo } from './resolvers/seo'
+import { SkuVariants } from './resolvers/skuVariations'
 import ChannelMarshal from './utils/channel'
 import type { Loaders } from './loaders'
 import type { Clients } from './clients'
@@ -80,6 +81,7 @@ const Resolvers = {
   StoreProductGroup,
   StoreSearchResult,
   StorePropertyValue,
+  SkuVariants,
   ObjectOrString,
   Query,
   Mutation,
